Consolidate Drude platform tests
