+++ plugins/drude/platforms/cuda/tests/CMakeLists.txt
@@ -6,6 +6,7 @@ ENABLE_TESTING()
 
 INCLUDE_DIRECTORIES(${CUDA_INCLUDE_DIR})
 INCLUDE_DIRECTORIES(${OPENMM_DIR}/plugins/drude/tests)
+INCLUDE_DIRECTORIES(${OPENMM_DIR}/platforms/cuda/tests)
 
 # Automatically create tests using files named "Test*.cpp"
 FILE(GLOB TEST_PROGS "*Test*.cpp")
