+++ plugins/drude/platforms/opencl/tests/TestOpenCLDrudeNoseHoover.cpp
@@ -29,32 +29,7 @@
  * USE OR OTHER DEALINGS IN THE SOFTWARE.                                     *
  * -------------------------------------------------------------------------- */
 
-//#include "ReferenceTests.h"
-
-#include "openmm/internal/AssertionUtilities.h"
-#include "openmm/Context.h"
-#include "openmm/NonbondedForce.h"
-#include "openmm/Platform.h"
-#include "openmm/System.h"
-#include "openmm/VerletIntegrator.h"
-#include "openmm/DrudeForce.h"
-#include "OpenCLPlatform.h"
-#include "SimTKOpenMMUtilities.h"
-#include <iostream>
-#include <vector>
-
-using namespace OpenMM;
-using namespace std;
-
-extern "C" OPENMM_EXPORT void registerDrudeOpenCLKernelFactories();
-
-void runPlatformTests() { }
-
+#include "OpenCLDrudeTests.h"
 #include "TestDrudeNoseHoover.h"
 
-Platform& initializePlatform(int argc, char* argv[]) {
-    registerDrudeOpenCLKernelFactories();
-    if (argc > 1) Platform::getPlatformByName("OpenCL").setPropertyDefaultValue("Precision", std::string(argv[1]));   
-    return Platform::getPlatformByName("OpenCL");
-}
-
+void runPlatformTests() {}
